requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-js-asset==1.2.2
django-markdown==0.8.4
django-mdeditor==0.1.20
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.1
install==1.3.5
Markdown==3.3.6
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.7.0
